getting user info from backend to fill form
