+++ package.json
@@ -10,12 +10,16 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
     "@testing-library/user-event": "^7.2.1",
+    "adaptiv-ui": "^0.6.5",
     "apollo-boost": "^0.4.7",
     "auth0-js": "^9.12.2",
-    "adaptiv-ui": "^0.6.5",
     "axios": "^0.19.2",
+    "from": "^0.1.7",
+    "gql": "^1.1.2",
     "graphql": "^14.6.0",
+    "graphql-tag": "^2.10.3",
     "react": "^16.12.0",
+    "react-apollo": "^3.1.3",
     "react-dom": "^16.12.0",
     "react-hook-form": "^4.9.8",
     "react-icons": "^3.9.0",
