+++ src/pages/DashRoute/queries/getProfile.js
@@ -1,4 +1,4 @@
-import { gql } from 'apollo-boost';
+import gql from 'graphql-tag';
 
 export const PROFILE_INFO = gql`
   query getProfile($email: String!) {
