--- src/pages/DashRoute/users/Profile.js
@@ -1,11 +1,0 @@
-import React from "react";
-import ProfileInfo from "./queries/ProfileInfo.js";
-
-export default function Profile() {
-  return (
-    <main>
-      <h1>Profile Component Working!</h1>
-      <ProfileInfo />
-    </main>
-  );
-}
